Remove unused string utility functions
These were holdovers from a previous PoC with iniparser.

The is_null_or_empty_check is technically not needed; there's an
argument that a user should be safeguarded from unexpected behavior in
the event that they're entering args of the form "" or '   ', but empty
strings aren't valid for any purposes relating to .ini files except
maybe as values.